introduce dev-libs/libsecp256k1-zkp
This is an experimental fork of libsecp256k1 by the Elements project to support Pedersen commitments and range proofs among other new features. This is needed for building net-libs/libwally-core with USE="elements".

The ebuild installs the library as libsecp256k1_zkp.so and the headers in /usr/include/secp256k1_zkp, so this forked version can coexist with dev-libs/libsecp256k1.